postman test for userCOntroller completed

+++ backend/OnlinebankingSystem/src/main/java/com/bank/OnlinebankingSystem/Repository/AccountDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountDao extends JpaRepository<com.bank.OnlinebankingSystem.Entity.Account,Integer> {
+public interface AccountDao extends JpaRepository<com.bank.OnlinebankingSystem.Entity.Account,Long> {
 
 }
